Updating build for OpenJDK 12

+++ package.sh
@@ -1,6 +1,7 @@
+#!/bin/bash -e
 JAVA=12
+JLINK=build/jdk/$JAVA/linux/bin/jlink
 
-#!/bin/bash
 git diff --exit-code preferences.properties > /dev/null
 if [ $? -eq 1 ]; then echo "Unclean preferences file."; exit; fi
 rm -rf "build/output/"
@@ -11,23 +12,22 @@ read -e -i "$version" -p "Edit version name: " version
 
 # javadoc
 echo "Generating Javadoc..."
-rm -r doc/javadoc
+if [ -d doc/javadoc ]; then rm -r doc/javadoc; fi
 javadoc -d doc/javadoc/ javelin  -subpackages javelin &>/dev/null
 
 #jlink
 function build() {
-    java=$1
-    system=$2
-    output="build/output/${system}/javelin"
-    echo "Building Javelin for $system (Java $java)..."
-    jlink --module-path .:build/jdk/${java}/${system}/jmods --add-modules javelin --output "$output/java"
-    cp -r build/launcher/$system/* doc avatars maps monsters.xml preferences.properties README.txt audio $output
-    echo $version>$output/doc/VERSION.txt
-    pushd $output/.. > /dev/null
-    zip "../javelin-${system}.zip" . -r > /dev/null
-    popd > /dev/null
-    rm -rf "build/output/${system}"
+	system=$1
+	output="build/output/${system}/javelin"
+	echo "Building Javelin for $system (Java $JAVA)..."
+	$JLINK --module-path .:build/jdk/${JAVA}/${system}/jmods --add-modules javelin --output "$output/java"
+	cp -r build/launcher/$system/* doc avatars maps monsters.xml preferences.properties README.txt audio $output
+	echo $version>$output/doc/VERSION.txt
+	pushd $output/.. > /dev/null
+	zip "../javelin-${system}.zip" . -r > /dev/null
+	popd > /dev/null
+	rm -rf "build/output/${system}"
 }
-build "$JAVA" "linux"
-build "$JAVA" "windows"
-build "$JAVA" "mac"
+build linux
+build windows
+build mac
